Refactor: replace UserMenuNew with UserMenu component
- Deleted the obsolete UserMenuNew component and updated Sidebar to import the UserMenu component instead. This change streamlines the authentication logic and improves code maintainability.
